this commit will delete setUser file and have them stored in the userContext
