+++ capsulert/contexts/User.js
@@ -1,10 +1,27 @@
-import { createContext, useState } from "react";
+import { createContext, useState, useEffect } from "react";
+import { getAuth, onAuthStateChanged } from "firebase/auth";
 
 export const UserContext = createContext();
 
 export const UserProvider = ({ children }) => {
   const [loggedInUser, setLoggedInUser] = useState("");
   const [userId, setUserId] = useState("");
+
+  const auth = getAuth();
+  useEffect(() => {
+    onAuthStateChanged(auth, (user) => {
+      if (user) {
+        //User is signed in
+        const uid = user.uid;
+        console.log("User uid is", uid);
+        setUserId(uid);
+        // ...
+      } else {
+        // User is signed out
+        setUserId("");
+      }
+    });
+  }, []);
 
   return (
     <UserContext.Provider
